WireGuard - fast, modern, secure kernel VPN tunnel
WireGuard is a novel VPN that runs inside the Linux Kernel and utilizes
**state-of-the-art cryptography**. It aims to be faster, simpler, leaner,
and more useful than IPSec, while avoiding the massive headache. It intends
to be considerably more performant than OpenVPN. WireGuard is designed as
a general purpose VPN for running on embedded interfaces and super computers
alike, fit for many different circumstances. It runs over UDP.

**More information may be found at [WireGuard.com](https://www.wireguard.com/).**

This project is released under the [GPLv2](COPYING).

Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>
Signed-off-by: nysascape <nysadev@raphielgang.org>